Rename ContentStore and associated variables
