+++ apps/main/web/templates/pages/about.html.slim
@@ -1,4 +1,4 @@
-- content_for.call(:title, "About")
+- content_store.capture(:title, "About")
 
 ul
   - people.each do |person|
