+++ apps/main/web/templates/pages/contact.html.slim
@@ -1,4 +1,4 @@
-- content_for.call(:title, "Contact")
+- content_store.capture(:title, "Contact")
 
 ul
   - contact_details.each do |contact_detail|
