+++ apps/main/web/templates/pages/home.html.slim
@@ -1,4 +1,4 @@
-- content_for.call(:title, "Home")
+- content_store.capture(:title, "Home")
 
 .page-intro.grid.grid--narrow.mx-auto
   h2.page-intro__heading
