+++ apps/main/web/templates/pages/notes.html.slim
@@ -1,4 +1,4 @@
-- content_for.call(:title, "Notes")
+- content_store.capture(:title, "Notes")
 
 .notes
   .pulled-heading.pulled-heading--no-underline.pulled-heading--mid.pulled-heading--color-blue.pulled-heading--border-blue.mx-auto
