+++ apps/main/web/templates/pages/work.html.slim
@@ -1,4 +1,4 @@
-- content_for.call(:title, "Work")
+- content_store.capture(:title, "Work")
 
 h1 Featured Items
 ul
