+++ apps/main/web/templates/posts/index.html.slim
@@ -1,4 +1,4 @@
-- content_for.call(:title, "Notes")
+- content_store.capture(:title, "Notes")
 
 == paginator.pager_links
